requirements.txt
Flask==2.3.2
python-telegram-bot==13.15
stripe==8.5.0